cockpit-config: widen dev_t in terminal-ingress identity (macOS)

+++ crates/cockpit-config/src/config/files.rs
@@ -3132,8 +3132,11 @@ pub(crate) fn rename_file_nofollow(source: &Path, destination: &Path) -> Result<
             );
         }
         let entry_identity = super::TerminalIngressFileIdentity {
-            volume: stat.st_dev,
-            file: stat.st_ino,
+            // `dev_t` is u64 on Linux but i32 on Darwin, so widen explicitly.
+            // This matches what `MetadataExt::dev()` does internally, keeping
+            // both sides of the identity comparison byte-identical.
+            volume: stat.st_dev as u64,
+            file: stat.st_ino as u64,
             links: stat.st_nlink.try_into().unwrap_or(u32::MAX),
         };
         if entry_identity != expected_identity {
